Added interfaces for services and @Repository annotations for repositories.
