+++ src/main/java/com/project/playlist/repository/PlaylistRepository.java
@@ -2,6 +2,8 @@ package com.project.playlist.repository;
 
 import com.project.playlist.model.Playlist;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface PlaylistRepository extends JpaRepository<Playlist, Long> {
 }
