+++ src/main/java/com/project/playlist/repository/VideoRepository.java
@@ -2,7 +2,9 @@ package com.project.playlist.repository;
 
 import com.project.playlist.model.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface VideoRepository extends JpaRepository<Video, Long> {
     public Video findByUrl(String url);
 }
